timestamps are now uint32 rather than float
Fixes #25.

+++ redditor.go
@@ -9,15 +9,15 @@ import (
 )
 
 type Redditor struct {
-	ID           string  `json:"id"`
-	Name         string  `json:"name"`
-	LinkKarma    int     `json:"link_karma"`
-	CommentKarma int     `json:"comment_karma"`
-	Created      float32 `json:"created_utc"`
-	Gold         bool    `json:"is_gold"`
-	Mod          bool    `json:"is_mod"`
-	Mail         *bool   `json:"has_mail"`
-	ModMail      *bool   `json:"has_mod_mail"`
+	ID           string `json:"id"`
+	Name         string `json:"name"`
+	LinkKarma    int    `json:"link_karma"`
+	CommentKarma int    `json:"comment_karma"`
+	Created      uint32 `json:"created_utc"`
+	Gold         bool   `json:"is_gold"`
+	Mod          bool   `json:"is_mod"`
+	Mail         *bool  `json:"has_mail"`
+	ModMail      *bool  `json:"has_mod_mail"`
 }
 
 // String returns the string representation of a reddit user.
